+++ lib/federails.rb
@@ -51,6 +51,19 @@ module Federails
       Configuration.actor_types[klass]
     end
 
+    # @return [Boolean] True if the given model is a possible data entity
+    def data_entity?(class_or_instance)
+      Configuration.data_types.key? class_or_instance_name(class_or_instance)
+    end
+
+    # @return [Hash] The configuration for the given data entity
+    def data_entity_configuration(class_or_instance)
+      klass = class_or_instance_name(class_or_instance)
+      raise "#{klass} is not a configured data entity" unless Configuration.data_types.key?(klass)
+
+      Configuration.data_types[klass]
+    end
+
     private
 
     # @return [String] Class name of the provided class or instance